vhost: let busy polling work at ns granularity

+++ drivers/vhost/vhost.h
@@ -148,7 +148,7 @@ struct vhost_virtqueue {
 	/* Ring endianness requested by userspace for cross-endian support. */
 	bool user_be;
 #endif
-	u32 busyloop_timeout;
+	u64 busyloop_timeout;
 };
 
 struct vhost_msg_node {
